requirements.txt
django-restframework
django-crispy-forms
django-widget-tweaks

